Make individual booking components for next and current/past
