+++ client/src/components/BookingItem/useStyles.ts
@@ -1,17 +1,21 @@
 import { makeStyles } from '@material-ui/core/styles';
 
 const useStyles = makeStyles((theme) => ({
+  main: {
+    display: 'flex',
+    justifyContent: 'space-between',
+  },
   pos: {
     marginBottom: 18,
+  },
+  details: {
+    display: 'flex',
+    alignItems: 'center',
   },
   large: {
     width: theme.spacing(7),
     height: theme.spacing(7),
     marginRight: theme.spacing(2),
-  },
-  details: {
-    display: 'flex',
-    alignItems: 'center',
   },
 }));
